seperate client form component and parent server page
